moved QR generation to CreateEvent page and named QR after event name
