+++ QRCodeAuth-Web/QRCodeAuth-Web/ManageEvent.aspx.cs
@@ -62,29 +62,7 @@ namespace QRCodeAuth_Web
 			lblStartTime.Text = string.Format("Start Time: {0}", ev.StartTime);
 			lblEndTime.Text = string.Format("End Time: {0}", ev.EndTime);
 			lblDescription.Text = string.Format("Description: {0}", ev.Description);
-
-			GenerateQR(ev);
-		}
-
-		protected void GenerateQR(Event ev)
-		{
-			// Convert string to event
-			string eventString = JsonConvert.SerializeObject(ev);
-
-			// get current path
-			string path = AppDomain.CurrentDomain.BaseDirectory;
-
-			// create barcode writer that creates QR Code
-			BarcodeWriter writer = new BarcodeWriter();
-			writer.Format = BarcodeFormat.QR_CODE;
-
-			// encode this text into the qr code and save it to the specified path 
-			writer.Write(eventString).Save(path + @"Images\QRCodes\generatedQR.jpg");
-
-			// load generated qr image into QR image control
-			eventQR.ImageUrl = path + @"Images\QRCodes\generatedQR.jpg";
-			eventQR.Visible = true;
-		}
+		}	
 
 		protected void GetLoggedInUserInfo()
 		{
