+++ QRCodeAuth-Web/QRCodeAuth-Web/Models/Event.cs
@@ -15,6 +15,7 @@ namespace QRCodeAuth_Web.Models
 		public string Name { get; set; }
 		public string Location { get; set; }
 		public EventType EventType { get; set; }
+		public DateTime Date { get; set; }
 		public DateTime StartTime { get; set; }
 		public DateTime EndTime { get; set; }
 		public string Description { get; set; }
